Imported free assets
But stayed up all night with two different tutorials (one of which was from last year!) and still can't figure out how to actually follow the tutorial.
Like I can't change the background image to a starfield. And now it's almost 6AM and I can't think straight anymore, and I have nothing to show for it.